requirements.txt
certifi==2017.4.17
chardet==3.0.4
click==6.7
cycler==0.10.0
dateutils==0.6.6
decorator==4.0.11
django-cors-headers==2.1.0
django-filter==1.0.4
Django==1.11.2
djangorestframework-jwt==1.11.0
djangorestframework==3.6.3
docopt==0.6.2
drf-extensions==0.3.1
ExifRead==2.1.2
face-recognition-models==0.1.3
face-recognition==0.1.14
Flask-Cors==3.0.3
Flask-RESTful==0.3.6
Flask==0.12.2
geopy==1.11.0
idna==2.5
ipdb==0.10.3
ipython-genutils==0.2.0
ipython==6.1.0
jedi==0.10.2
lxml==4.1.1
matplotlib==2.0.2
networkx==1.11
nltk==3.2.4
numpy==1.13.0
olefile==0.44
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
pipreqs==0.4.7
prompt-toolkit==1.0.14
psycopg2==2.7.1 # for postgresql
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
python-memcached==1.58 # for memcached
pytz==2017.2
requests==2.18.1
scikit-learn==0.18.2
scipy==0.19.1
simplegeneric==0.8.1
six==1.10.0
tornado==4.5.1
tqdm==4.14.0
traitlets==4.3.2
vobject==0.9.5
urllib3==1.21.1
wcwidth==0.1.7
yarg==0.1.9
